Allow to limit the number of concurrent streams per http/2 connection and use a connection pool

+++ src/main/java/io/vertx/core/http/impl/Http2ClientConnection.java
@@ -54,7 +54,7 @@ class Http2ClientConnection extends Http2ConnectionBase implements HttpClientCon
   final Http2Pool http2Pool;
   final HttpClientMetrics metrics;
   final Object metric;
-  long streamCount;
+  int streamCount;
 
   public Http2ClientConnection(Http2Pool http2Pool,
                                ContextImpl context,
